requirements.txt
google-genai
python-dotenv
requests
fastapi
uvicorn
pydantic
